refactor: move mutate process from transaction to bloctowalletprovider
